global error handler with route specific error is configured
